Project: Replace "Fixes" with "Issue" in PR template
We ran into a problem with #2593 where it was improperly resolved.
PR #3328 did not fully resolve the issue, but was merged because it made
necessary improvements. However, because the PR template contained
"Fixes #...", when PR #3328 was merged GitHub automatically closed
Issue #2593. This flew under the radar and Issue #2593 was therefore not
properly resolved until 2 releases later.

By changing this template, we hope to avoid similar sequences of events
in the future.