Docs: Remove dates and weights
Hugo orders items in the navigation bar first by weight, then by date.
Since this isn't a blog, the date is not useful metadata to keep in the
file; that's what we have version control (git) for. That leaves weight,
which is somewhat inconvenient. In most cases, we want things sorted by
title and only sorted by weight if weight is present. To that end, the
first step is removing weights from all of the files.

Hugo (and the learn theme) doesn't actually support sorting by weight
and name. As we add weights back in to the top-level items, we may also
have to modify the learn theme in order to get ordering the way we want
it.